rename: 모듈명 변경 (통일성)
